+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
+### Added
+* `Copy`, `Clone` and `Hash` on error & event types (where possible)
 ### Changed
 * The MSRV has been updated to 1.81.0 due to `core::error::Error` being implemented
 * **BREAKING**: the features `use_alloc` and `use_heapless` have been renamed to `alloc` and `heapless` respectively.